app/backend/requirements.txt
Flask
psycopg2-binary
Flask-CORS
python-ldap
pyjwt
flask-jwt-extended